pyproject.toml
[tool.poetry]
name = "example-fs-ds-repo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "example_fs_ds_repo"}]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
